[bedrock] Teach Dict to TryIntoOpen
Previously the Dict capability did not support conversion into Open. It
turns out with a few lines of logic we can convert a Dict to Open:

- Recursively convert all members to Open
- Put those members in a vfs directory
- Call `dir.clone().open()` in the OpenFn

This sets us up for the next step which is to use it in
serve_processargs.

Change-Id: I17f6e32d7e28e827636eca4e6901286f5e8faf2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/880986
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>