[driver_runtime] Correctly handle no signal packet
The signal packet may be nullptr when the wait is canceled.
Dereferencing it causes a crash, so instead we store a std::optional<>
signal packet to remember if we recieved a valid signal packet to begin
with.

Change-Id: I9c1d6d2b247a29aca30f7e0ba0b867a1068aec90
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883812
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>