[libc] Temporarily disable ubsan on find_buildid_note
bringup.riscv64-asan is crashing with a stackoverflow in this function.
This function in particular uses a lot of stack for
-fsanitize=null,object-size,alignment checks. UBSan adds calls to
__ubsan_handle_type_mismatch_v1 which, on the failing case, accepts
pointers to globals containing type information. This function does a
lot of these checks and decides to store all these pointers onto the
stack at the start of the function which leads to the overflow.

In the long run, the new libc/dynamic linker may address some of this
and the instrumentation story will be different, but in the meantime, we
can selectively disable ubsan instrumentation on some early startup
code.

At the moment, it looks like just disabling ubsan on find_buildid_node
only seems to get us part the startup phase. I figure we can keep the
macro in case we may need to add it elsewhere in the future if a new
clang roll leads to different codegen which may use more stack.

Bug: 130415, 130526
Change-Id: I9103a15d3d0c0a1fc5187ecae4dac0e100615006
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883574
Commit-Queue: Leonard Chan <leonardchan@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>