[rbe] Suppress writing stub on matching output
This accounts for interactions between the rewrapper features
--download_outputs=false and --preserve_unchanged_output_mtime.

When a download is skipped, it could be attributed to either
of these features, but download stub creation should depend
on which case it is.

If an output already exists and is valid (its digest matches
that reported by the corresponding output_file in the action_log),
we want to leave it as-is, without overwriting it with a download stub.
If the existing output is a download stub, we update it only if the
blob_digest changes.

This lets us continue to have the benefit of leaving timestamps
untouched where it is correct to do so (for ninja restat).

Bug: 123178
Change-Id: I61136c29fc7aaaf48e4d87bcea53654862a60ba1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864305
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Fang <fangism@google.com>