[scenic][flake] Capture a weak ptr instead of this
It seems that capturing `this` in this specific closure can make the
this pointer outlive the actual object.

The simplest way to remove that issue is to use a weak ptr instead,
but it is not super-obvious that no other issues lurk here.
Specifically, `Endpoint` destruction is nontrivial.
It still seems like an adequate thing to do.

Bug: 128914
Change-Id: Ib8e2c7f69e16cb1ef82e67cf6b28d3fbb5f2eb8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883533
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>