[fdf] Don't return host koid when unbound
Currently when the node is unbound we can still return a host koid. This
is because the driver host pointer being used is a raw pointer into the
self deleting driver host list entry. It doesn't make sense to have a
host koid for an unbound driver, so now we check to make sure the node
is bound before trying to find the host koid.

Change-Id: I51f91943fe71c0068feb3c04cdd1f68d736fc82c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883717
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>