[async-loop] Fix race between PostTask and Shutdown
If a PostTask was running and Shutdown is called at the same time, the
PostTask could succeed in inserting its task into the list after the
shutdown finished. If that happens, then the task will never be run (not
even with ZX_ERR_CANCELED) which could cause problems. For example, with
an async::PostTask the memory allocated for the task will be leaked.

Multiply: async-loop-test-pkg
Change-Id: Id901f0110074a47a8f743d26034b32c1af283a2f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884075
Commit-Queue: John Bauman <jbauman@google.com>
Reviewed-by: John Grossman <johngro@google.com>