[ld] Refactor bootstrapping
This refactors bootstrapping so that the ld::abi Module data
structures for the vDSO and for the dynamic linker itself are
filled out along the way.

It also cleans up the build details a bit, renaming all the
targets to be called ld-* instead of *-ld. The ld-startup target
is now a shared_library, so it will use the user.basic variant
corresponding to the variant of its dependent (test).  The files
like this used by the test but not linked into the test program
go in data_deps, not deps. This matters when they include
linkable targets like a shared_library.

Bug: 130483
Change-Id: Icd58d060c3ddfc86a64bc2742285436f6079859d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883513
Reviewed-by: Alex Brachet <abrachet@google.com>
Commit-Queue: Roland McGrath <mcgrathr@google.com>