[ffx][fastboot] reland boot timeout success
(This is a reland of fxr/882214 -- the problem was that a
usb_bulk::read() will always timeout after 800ms; when running
a read with a longer timeout, the code needs to retry.)

When issuing a "boot", "continue", "reboot", or "reboot-bootloader",
it's possible that the "okay" response gets lost in a buffer. So inside
of consdering a lack of an "okay" in those cases as a a timeout failure,
consider them successes.

Also, since we are now considering a timeout to be an expected
result, let's timeout in only a few seconds instead of waiting
30 seconds.

This change includes a fix to fastboot::read_with_timeout() to
actually use a timeout, rather than blocking and checking the
timestamp after the fact.  Note that since usb_bulk will timeout
after 800ms, we have to retry if we get ETIMEDOUT.

Change-Id: I46cf8a83c1c91e26cb539206325e864e320c14c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883118
Commit-Queue: Steven Grady <slgrady@google.com>
Reviewed-by: Dwayne Slater <dwayneslater@google.com>