[magma] Don't use shared_ptr with MagmaSystemDevice
When creating a device, the driver framework gives the device temporary
ownership of certain hardware resources (e.g. interrupts). Those
resources must released by the driver before driver teardown completes.

When MagmaSystemDevice is held as a shared_ptr by the
MagmaSystemConnection, the lock on it in the MagmaSystemConnection
destructor may hold on to it after MagmaSystemDevice::Shutdown() is
called, which may cause the resources to not be released in a timely
manner.

Since the MagmaSystemDevice always outlives the MagmaSystemConnection
anyway (because Shutdown() tears down the connection), we don't need the
weak_ptr in MagmaSystemConnection; removing it fixes the lifetime issue.

Verifying that the MagmaSystemConnection outlives the MagmaSystemDevice
is difficult with the current way we use shared_ptrs with
PrimaryFidlServer. To fix that and also ensure that the FIDL bindings
are being used in a thread-safe way, add a PrimaryFidlServerHolder class
that's responsible for ensuring that the PrimaryFidlServer is torn down
on the correct thread; PrimaryFidlServerHolder can have external
references taken to it and can have an arbitrarily-long lifetime.

Multiply: magma_sys_driver_test
Bug: 129953
Change-Id: If8ef650bd7c2e70665f4d6b7be624c14b0c1b166
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/883133
Reviewed-by: Craig Stout <cstout@google.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>