[bt][obex] Send & Receive the Connection ID
The Connection ID is used to uniquely identify an OBEX session between a
Client and Server. It is optionally determined in the CONNECT operation,
and if provided by the Server, is used for the lifetime of the session.

The ID is uniquely generated by the OBEX Server and must be included in
the first packet of every subsequent operation after the CONNECT.

Bug: 128855
Test: fx test bt-obex-tests
Change-Id: I4b019b8525985d0d9fc51d9a29d00a8089e3f04b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/871292
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>