[ld] Wire up allocators
This sets up the //src/lib/trivial-allocator machinery for use in
the dynamic linker.  There is a scratch allocator for ephemeral
uses during startup, and there is an initial-exec allocator that
will be used to populate the passive ABI.

For the Fuchsia incarnation's tests, this now plumbs the "root"
VMAR through processargs.  The in-process tests use a 1GiB VMAR
to isolate allocations, and destroy it at the end of the test.

Bug: 130483
Change-Id: Id953d1446d56d62be2ef498002a84a7081566879
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/885071
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>