[bt][obex] Encode headers in order
Use a BTreeMap to store headers in the HeaderSet collection. This
provides for efficient insertion and removal, and fairly efficient
lookup. The size of this collection is bounded by the number of unique
HeaderIdentifiers (26 + user defined). Prefer BTreeMap over HashMap due
to the ordering invariant that exists for the BTreeMap. By maintaining
an (arbitrary) ordering, verifying correctness in unittests is
simplified as the encoded buffer will be encoded in the same order each
time. With a HashMap, encoding ordering is not guaranteed and makes
asserting data validity difficult.

Add TODOs for future ordering constraints when the relevant features are
added.

Bug: 125070
Test: fx test bt-obex-tests
Change-Id: I0254ceabd7e7426c28fc112161b79f247f2fd38a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872722
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Fuchsia-Auto-Submit: Ani Ramakrishnan <aniramakri@google.com>