[emulator_test] Disallow global owners
This test is designed to block large scale changes in fuchsia.git that
might break the Netstack team's out of tree conformance testing.
Disallow global owners on the test source so that a Netstack team member
will have to be included when the test needs to be revised (therefore
signaling a potential breakage).

Change-Id: Ic2710af4075d1f5ba244c6a3e0667fc2466a242f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884159
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Nick Brown <nickbrow@google.com>
Reviewed-by: Gabe Schine <thatguy@google.com>