[socket-tun] Send message with varied bytes
As a result of discovering and fixing https://fxbug.dev/128850, change
the bytes that are sent in this benchmark to something more varied than
a single byte repeated ad infinitum, because any permutation of such a
message is equal to the original message and a sender asserting that
the received bytes is equal to the sent bytes cannot tell that the
message was in fact sent out-of-order.

Add assertions that sent bytes is equal to received bytes.

Always set the socket option to modify the send buffer size (this was
not previously done because lowering the send buffer size to 10KiB was
what tickled the issue).

Test: Manually verified that tun-socket-benchmarks-ns3-with-tracing
      produces traces that do not exhibit the symptoms fixed in
      https://fxbug.dev/128850.
Change-Id: I4d9729dbff382e543c66b452910324b18721e10e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/874976
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>