[superproject] Roll llvm-project/libc [libc] Treat the locks array as a bitfield
Currently we keep an internal buffer of device memory that is used to
indicate ownership of a port. Since we only use this as a single bit we
can simply turn this into a bitfield. I did this manually rather than
having a separate type as we need very special handling of the masks
used to interact with the locks.

Reviewed By: JonChesterfield

Differential Revision: https://reviews.llvm.org/D155511

GitOrigin-RevId: 268e65cda60532cc422c948b671f03b58c0e2c72
Original-Revision: ffdf5b2524f542eadcfd302f73914b2a4c14157e
Roller-URL: https://ci.chromium.org/b/8775196129381172001
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ifec9bf5b6505dd02cd08f6c6352b382676d4ebc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886614