[starnix] Split filesystem creation between CurrentTask and Kernel
Filesystems can be created from a Kernel object and an external resource
or they can be specific to the currently executing task. FUSE
filesystems and ext4 filesystems backed by open files are examples of
the latter. This splits the filesystem creation routines based on the
caller's context so that filesystems that are not task specific can be
created without a CurrentTask reference.

Bug: 130182

Change-Id: Icb8c3af184bf0db7903928988a17ec3a36790fae
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/885633
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>