[fxtest][rewrite] Implement asynchronous command execution.
asyncio support spawning and monitoring subprocesses asynchronously, but
the official docs include an ominous warning about awaiting on output
streams directly. This is because awaiting, for instance, stderr while
there is data waiting to be written to stdout could cause the writing
program to hang (if the pipe is full and cannot take more data). In this
case, the reading program waits indefinitely for more stderr (which will
never come) while the writing program hangs indefinitely for there to be
space to write in stdout (which will never happen).

To solve this problem, this CL provides AsyncCommand, which handles
spawning individual async tasks to read stderr and stdout concurrently.
Because asyncio internally connects these pipes using non-blocking
async readers, we do not hang.

For convenience, AsyncCommand supports the following:
- Async iteration over output and termination events for wrapped
  commands.
- Line splitting.
- Environment handling, including changing working directories.
- Runtime calculation.
- Output parsing safely into UTF-8.
- Piping output through another program (to be used for stack trace
  symbolization).

Change-Id: I1e489d7d8530e75f969435728b2606e08e6208ed
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884202
Commit-Queue: Christopher Johnson <crjohns@google.com>
Reviewed-by: Kevin Cho <chok@google.com>