[bt][hfp] Move call types to a library.
This change move the call types that we expect to use in the HFP
Hands-Free component out of the HFP Audio Gateway and into a library to
be used by both components.  We expect to add more types to the library
crate, bt-hfp, as HFP Hands-Free is developed, and the name reflects
that.

Change-Id: I21f80f3001aeaae3b2798fb10b9b59d13d5517d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872540
Reviewed-by: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>
Commit-Queue: Matthew Kehrt <kehrt@google.com>