[clang-tidy] Disable misc-include-cleaner
This CL disables misc-include-cleaner because it is disrupting people's
workflow with large numbers of Tricium comments on C++ CLs, some of
which appear to be wrong or contradictory. The rule may be re-enabled in
the future after problems are worked out and after landing a mass change
to make existing code comply with it.

Change-Id: I68cd761d7388a1c108031577d0ffe7d60343544d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887251
Fuchsia-Auto-Submit: Mitchell Kember <mkember@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>