[Py] Update python build and test configs from 3.8 to 3.11
This commit updates the python GN build and test configurations to
use Python 3.11.
* The vendored Python3.11 bootstrap script has been added.
* The generic version symlink now points to 3.11.
* The python GN build configs have been updated from 3.8 to 3.11.

In addition,this commit removes manual scope parsing for IPv6 addresses.
In Python 3.11, the ipaddress module now automatically parses the scope
for IPv6 addresses. As a result, we no longer need to manually parse
the scope after upgrading to Python 3.11.

We decided to keep the Python 3.8 version to make rollback simpler,
in case of failures. Once we have upgraded all of the Python 3.8 pins
to Python 3.11, we will remove the Python 3.8 prebuilt version.

BUG=b/280829652

Change-Id: I868e7c0f008af1b251cf95ae6fe2b55fb02b0ad0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/864776
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Commit-Queue: Mahesh Saripalli <maheshsr@google.com>