[boot-shim] Introduce CPU topology decoding.
While specific for RISC-V, most of the complexities are shared with ARM
and will be generalized once RISCV support is complete.

This introduces an item capable of generating the CPU topology data from
the devicetree.

For performance class, all clusters have a default performance class of
1, unless a CPU specifies scaling property wihin that cluster. At that
point the highest capacity is picked as a representative of the cluster.

Then the performance is scaled down to 0-255 range where 255 is the
performance class of the cluster with the highest performance class.

Performance class is meant to be a relative comparison between CPUs, for
scheduling purposes.

Also the devicetree boot shim now appends the CPU topology for riscv.

Bug: 114124
Change-Id: I23772b55d93738615f48b6b5fbce3b293a34fd67
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879312
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>