[amlogic_decoder] Implement driver release hook
This was historically not necessary to implement because we never call
the releas hook in DFv1. In DFv2, the compat driver will actually call
it when the number of driver instances for a particular driver hits 0.
This happens during the amlogic decoder integration tests which unload
the driver in favor of another test driver.

Change-Id: I7225cdae4b15dc96783d5a719ace16ae5e6071d1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887229
Reviewed-by: Nick Eaton <nickeaton@google.com>
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>