[assembly] BoardInformation is no longer optional
Require BoardInformation to be specified for all assembled systems.

This required also ensuring that any host-test that runs assembly, and
uses its output as test data, isn't included in the build when a board
is not defined (sdk builders, and host-test-only builders are in this
class of configurations).

Also moves scrutiny tests into the tree of 'host_tests' groups, and
only defines those tests when a board has been defined for the build.

Bug: b/287519036

Change-Id: I5956f0057c43f63f8b63f1a352113127053f382e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/886659
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: David Iserovich <diserovich@google.com>
Commit-Queue: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>