[assembly] Deprecate fvm_emmc/fvm_ftl gn args
These two args:
- fvm_emmc_partition_size
- fvm_ftl_nand_block_count

Are only used in assembled_system() to determine whether to declare the
optional fastboot and nand fvms as outputs. We are replacing these two
arguments with ones that are better named.

We are also using this opportunity to remove the usage of a few global
gn arguments in assembled_system() and forcing clients to specify the
images they expect to be generated:
- generate_fvm
- generate_fvm_fastboot
- generate_fvm_nand
- generate_fxfs
- generate_vbmeta

Bug: b/291945693
Change-Id: Iffb73730db7eba31a3286a554c7393ce8d9b5dca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/887234
Reviewed-by: James Sullivan <jfsulliv@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>