[rust][vfs] provide path, not name, to not found handler
A simple directory in the Rust VFS may have a handler registered with
it, which is called whenever an open request is received for a path that
the VFS is not configured to handle.

That handler should be called with the full path that cannot be served,
instead of the single node in the path that the not found issue was
caused by.

Change-Id: I953a1d2d62c1a4be2a60924df403cd2f6cf1042a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/879904
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>